% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSfrags}
\alias{MSfrags}
\title{MSfrags
Sample peptide fragments in Spectonaut format}
\format{
A character vector containing 4 sample peptide fragments
with protein modifications in Spectronaut format
}
\source{
These fragments were generated from P51681
(for details, see ...)
}
\usage{
MSfrags
}
\description{
MSfrags
Sample peptide fragments in Spectonaut format
}
\keyword{datasets}
